Webpage around wasm
I'll try to keep the wasm_exec.js up to date with the version of Go I'm using (yes, it apparently DOES change).
index.html started out as the default one that comes with Go, but this has been drastically updated to wire the canvas to the wasm.
I instinctively put the dist folder in .gitignore but there's custom dev work here. So I'm adding them. They're the same ones from the, now deleted, goboy-wasm repo.