Update tests for new setup
Updated a path for a test image. It's not a file I can check in (I think? I'm playing it safe) as it's the "bios" startup code that initializes memory, drops the Nintendo logo, and plays the 2 note jingle.
Fixed test env variables so tests can be run. Wasm is the target build for "prod" but testing can run natively.