Tested build scripts on mac, needed tweaking
Running `go generate ./...` on a Mac resulted in "GOOS=js is an unrecognized command" and also tried to run the script in the windows file.
Added proper  build tags to the windows file and Mac stopped trying to run it.
Changed up how Mac/Linux builds by opening a shell and that shell parsing the command correctly (exact same approach as windows).

Need to check on my linux laptop next...