More progress
Now using the correct startup rom. Use DMG_ROM.bin instead of dmg0_rom.bin. A link to it has been added in the README. Not sure what I had before but instruction $0021 in it was `LD HL, $0104` instead of `LD DE, $0104` as it should be.

Moving forward I'm indicating exactly what variant of an instruction is being handled in opcodes.go in a comment.

First attempt at implementing the stack. The key is that it's 16 bits wide, PUSHing decrements the stack pointer, and POPing increments it because it should point to the last byte in memory when ROM execution first begins.

No longer panic with strings.

Assert library in tests.

More instructions implemented.

Restructured implementations around table reads and writes.