Touch Ups
Touch ups like setting curGain at init, allowed the disabling of FPS, noticed that all the work I was doing to "get pixel data" is already done for me, and migrating a defer to the end of the function. These changes have improved the frame draw time (time it takes to transfer a complete frame from wasm/go to the canvas) from ~1.1ms to ~0.45ms.

Updated the go version to current.

Added a just file.

Updated libraries.